Added all of the sprites from the intro scene
I also measured the timestamps needed in order for the animation to work, but I haven't found a way to put them into a viewable file. It's very possible that the cutscene will have to be done in colaboration with me, so when someone is ready to start coding let me know and I'll fill you in on the info and the structure I've prepared for the cutscene.